requirements.txt
# --- Podstawowe zależności ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx>=0.25.2,<0.26.0
psycopg2-binary==2.9.9
greenlet==3.0.3
streamlit==1.30.0

# --- Baza i Cache ---
asyncpg==0.29.0
redis==5.0.1
alembic==1.13.1

# --- GEN-AI Stack (Obsługa OpenAI i Ollama) ---
openai==1.10.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.13
ollama==0.1.6

# --- Security & Auth (JWT/JOSE) ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9